svn merge --reintegrate svn://macaulay2.math.uiuc.edu/Macaulay2/branches/dan-mike/2010January



git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@11223 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/e/buffer.hpp
@@ -41,7 +41,7 @@ public:
   M2_string to_string(); // Copies the string, leaves buffer intact
 
   void put(const char *s);		// Place null-terminated string into buffer
-  void put(const char *s, int len);	// Place a string possible containing null chars
+  void put(const char *s, long len);	// Place a string possible containing null chars
   void put(char c);		// Place a single character
   void put(int n);		// Format the integer, place into buffer
   void put(int n, int width);	// Format the integer, with given width field.
