+++ Macaulay2/e/dmat.hpp
@@ -18,17 +18,17 @@ class DMat : public our_new_delete
   typedef typename CoeffRing::ring_type RingType;
   const RingType *R; // To interface to the outside world
   CoeffRing * coeffR; // Same as R, optimized for speed.  R->get_CoeffRing()
-  long nrows_;
-  long ncols_;
+  int nrows_;
+  int ncols_;
   elem *array_; // array has length nrows*ncols
                 // columns stored one after another
 
 
-  void copy_elems(long n_to_copy, elem *target, long target_stride, elem *source, long stride);
+  void copy_elems(long n_to_copy, elem *target, int target_stride, elem *source, int stride);
 public:
   DMat():R(0), coeffR(0), nrows_(0), ncols_(0), array_(0) {} // Makes a zero matrix
 
-  DMat(const RingType *R0, long nrows, long ncols); // Makes a zero matrix
+  DMat(const RingType *R0, int nrows, int ncols); // Makes a zero matrix
 
   void grab(DMat *M);// swaps M and this.
 
@@ -36,14 +36,14 @@ public:
 
   bool is_dense() const { return true; }
 
-  long n_rows() const { return nrows_; }
-  long n_cols() const { return ncols_; }
+  int n_rows() const { return nrows_; }
+  int n_cols() const { return ncols_; }
   const RingType * get_ring() const { return R; }
   const CoeffRing * get_CoeffRing() const { return coeffR; }
 
   void set_matrix(const DMat<CoeffRing> *mat0);
-  void initialize(long nrows, long ncols, elem *array);
-  void resize(long nrows, long ncols);
+  void initialize(int nrows, int ncols, elem *array);
+  void resize(int nrows, int ncols);
   elem * get_array() const { return array_; } // Used for lapack type routines
 
   double *get_lapack_array() const; // redefined by RR,CC
@@ -54,7 +54,7 @@ public:
   class iterator : public our_new_delete
   {
     const DMat<CoeffRing> *M;
-    long col;
+    int col;
     elem *begin;
     elem *end;
     void to_next_valid() { 
@@ -63,7 +63,7 @@ public:
       while (begin >= end && R->is_zero(*begin)) --begin;
     }
   public:
-    void set(long col0) { 
+    void set(int col0) { 
       col = col0; 
       begin = M->array_ + (col0+1) * (M->n_rows()); 
       end = begin-M->n_rows();
@@ -76,7 +76,7 @@ public:
     const elem &value() { return *begin; }
     void next() { to_next_valid(); }
     bool valid() { return begin >= end; }
-    long row() { return begin-end; }
+    int row() { return static_cast<int>(begin-end); }
 
     void copy_elem(ring_elem &result) { 
       M->get_CoeffRing()->to_ring_elem(result, value());
@@ -85,11 +85,11 @@ public:
 
 
 public:
-  long lead_row(long col) const;
+  int lead_row(int col) const;
   /* returns the largest index row which has a non-zero value in column 'col'.
      returns -1 if the column is 0 */
 
-  long lead_row(long col, elem &result) const;
+  int lead_row(int col, elem &result) const;
   /* returns the largest index row which has a non-zero value in column 'col'.
      Also sets result to be the entry at this index.
      returns -1 if the column is 0, or if col is out of range
@@ -101,67 +101,67 @@ public:
   // The following routines return false if one of the row or columns given
   // is out of range.
 
-  bool get_entry(long r, long c, elem &result) const;
+  bool get_entry(int r, int c, elem &result) const;
   // Returns false if (r,c) is out of range or if result is 0.  No error
   // is returned. result <-- this(r,c), and is set to zero if false is returned.
 
-  void set_entry(long r, long c, const elem &a);
+  void set_entry(int r, int c, const elem &a);
   // Returns false if (r,c) is out of range, or the ring of a is wrong.
 
-  void interchange_rows(long i, long j); 
+  void interchange_rows(int i, int j); 
   /* swap rows: row(i) <--> row(j) */
 
-  void interchange_columns(long i, long j); 
+  void interchange_columns(int i, int j); 
   /* swap columns: column(i) <--> column(j) */
 
-  void scale_row(long i, const elem &r); 
+  void scale_row(int i, const elem &r); 
   /* row(i) <- r * row(i) */
 
-  void scale_column(long i, const elem &r); 
+  void scale_column(int i, const elem &r); 
   /* column(i) <- r * column(i) */
 
-  void divide_row(long i, const elem &r);
+  void divide_row(int i, const elem &r);
   /* row(i) <- row(i) / r */
 
-  void divide_column(long i, const elem &r);
+  void divide_column(int i, const elem &r);
   /* column(i) <- column(i) / r */
 
-  void row_op(long i, const elem &r, long j); 
+  void row_op(int i, const elem &r, int j); 
   /* row(i) <- row(i) + r * row(j) */
 
-  void column_op(long i, const elem &r, long j); 
+  void column_op(int i, const elem &r, int j); 
   /* column(i) <- column(i) + r * column(j) */
 
-  void column2by2(long c1, long c2, 
+  void column2by2(int c1, int c2, 
 		  const elem &a1, const elem &a2,
 		  const elem &b1, const elem &b2);
   /* column(c1) <- a1 * column(c1) + a2 * column(c2),
      column(c2) <- b1 * column(c1) + b2 * column(c2)
   */
 
-  void row2by2(long r1, long r2, 
+  void row2by2(int r1, int r2, 
 	       const elem &a1, const elem &a2,
 	       const elem &b1, const elem &b2);
   /* row(r1) <- a1 * row(r1) + a2 * row(r2),
      row(r2) <- b1 * row(r1) + b2 * row(r2)
   */
 
-  void dot_product(long i, long j, elem &result) const; 
+  void dot_product(int i, int j, elem &result) const; 
 
-  bool row_permute(long start_row, M2_arrayint perm);
+  bool row_permute(int start_row, M2_arrayint perm);
 
-  bool column_permute(long start_col, M2_arrayint perm);
+  bool column_permute(int start_col, M2_arrayint perm);
 
-  void insert_columns(long i, long n_to_add);
+  void insert_columns(int i, int n_to_add);
   /* Insert n_to_add columns directly BEFORE column i. */
 
-  void insert_rows(long i, long n_to_add);
+  void insert_rows(int i, int n_to_add);
   /* Insert n_to_add rows directly BEFORE row i. */
 
-  void delete_columns(long i, long j);
+  void delete_columns(int i, int j);
   /* Delete columns i .. j from M */
 
-  void delete_rows(long i, long j);
+  void delete_rows(int i, int j);
   /* Delete rows i .. j from M */
 
   bool set_submatrix(M2_arrayint rows,
